# astroneko-backend/docs/swagger.yaml
basePath: /
definitions:
...
  referral_code.GetUserReferralCodesResponse:
    properties:
      codes:
        items:
          $ref: '#/definitions/referral_code.UserReferralCodeResponse'
        type: array
    type: object
...
  referral_code.UserReferralCodeResponse:
    properties:
      id:
        type: string
      is_activated:
        type: boolean
      referral_code:
        type: string
    type: object
  shared.ResponseBody:
    properties:
      currentPage:
        type: integer
      data: {}
      latestAiSuggestQuestions:
        items:
          type: string
        type: array
      meta:
        additionalProperties: true
        type: object
      perPage:
        type: integer
      status:
        $ref: '#/definitions/shared.Status'
      totalItem:
        type: integer
    type: object
  shared.Status:
    properties:
      code:
        type: string
      errorID:
        type: string
      message:
        items:
          type: string
        type: array
      successID:
        type: string
    type: object
  user.ActivateReferralRequest:
    properties:
      referral_code:
        type: string
    required:
    - referral_code
    type: object
  user.ActivateReferralResponse:
    properties:
      currentPage:
        type: integer
      data: {}
      latestAiSuggestQuestions:
        items:
          type: string
        type: array
      meta:
        additionalProperties: true
        type: object
      perPage:
        type: integer
      status:
        $ref: '#/definitions/shared.Status'
      totalItem:
        type: integer
    type: object
...
  user.DeleteUserResponse:
    properties:
      currentPage:
        type: integer
      data: {}
      latestAiSuggestQuestions:
        items:
          type: string
        type: array
      meta:
        additionalProperties: true
        type: object
      perPage:
        type: integer
      status:
        $ref: '#/definitions/shared.Status'
      totalItem:
        type: integer
    type: object
...
  user.GetUserResponse:
    properties:
      currentPage:
        type: integer
      data: {}
      latestAiSuggestQuestions:
        items:
          type: string
        type: array
      meta:
        additionalProperties: true
        type: object
      perPage:
        type: integer
      status:
        $ref: '#/definitions/shared.Status'
      totalItem:
        type: integer
    type: object
...
  user.UpdateUserRequest:
    properties:
      display_name:
        type: string
      is_activated_referral:
        type: boolean
      latest_login_at:
        type: string
      profile_image_url:
        type: string
    type: object
  user.UpdateUserResponse:
    properties:
      currentPage:
        type: integer
      data: {}
      latestAiSuggestQuestions:
        items:
          type: string
        type: array
      meta:
        additionalProperties: true
        type: object
      perPage:
        type: integer
      status:
        $ref: '#/definitions/shared.Status'
      totalItem:
        type: integer
    type: object
...
  user_limit.GetUserLimitResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      limit:
        type: integer
      updated_at:
        type: string
    type: object
  user_limit.IsUserOverLimitUsedResponse:
    properties:
      is_over_limit:
        type: boolean
    type: object
  user_limit.UpdateUserLimitRequest:
    properties:
      limit:
        minimum: 300
        type: integer
    required:
    - limit
    type: object
  user_limit.UpdateUserLimitResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      limit:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for Astroneko Backend
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Astroneko Backend API
  version: "1.0"
paths:
...
  /v1/api/crm/user-limit:
    put:
      consumes:
      - application/json
      description: Update the user limit configuration (CRM access required)
      parameters:
      - description: Updated user limit data
        in: body
        name: user_limit
        required: true
        schema:
          $ref: '#/definitions/user_limit.UpdateUserLimitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_limit.UpdateUserLimitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      security:
      - BearerAuth: []
      summary: Update user limit
      tags:
      - user-limit
...
  /v1/api/user-limit:
    get:
      consumes:
      - application/json
      description: Get the current user limit configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_limit.GetUserLimitResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      summary: Get user limit
      tags:
      - user-limit
  /v1/api/user-limit/check:
    get:
      consumes:
      - application/json
      description: Check if the total number of users exceeds the configured limit
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_limit.IsUserOverLimitUsedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      security:
      - BearerAuth: []
      summary: Check if user limit is exceeded
      tags:
      - user-limit
...
  /v1/api/users/{id}/referral/activate:
    post:
      consumes:
      - application/json
      description: Activate a referral code for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Referral data
        in: body
        name: referral
        required: true
        schema:
          $ref: '#/definitions/user.ActivateReferralRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.ActivateReferralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      summary: Activate referral code
      tags:
      - users/referral
  /v1/api/users/{id}/referral/codes:
    get:
      consumes:
      - application/json
      description: Get all referral codes for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/referral_code.GetUserReferralCodesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      summary: Get user referral codes
      tags:
      - users/referral
...
  /v1/api/users/profile/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.DeleteUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      summary: Delete user
      tags:
      - users/profile
    get:
      consumes:
      - application/json
      description: Get user information by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.GetUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      summary: Get user by ID
      tags:
      - users/profile
    put:
      consumes:
      - application/json
      description: Update user information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UpdateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      summary: Update user
      tags:
      - users/profile
  /v1/api/users/profile/me:
    get:
      consumes:
      - application/json
      description: Get current authenticated user information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.GetUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users/profile
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"